small updates for brevity

+++ canvas/canvas_count_pdf_pages.py
@@ -12,9 +12,16 @@ canvasToken = realm["canvasToken"]
 canvasApi = realm["canvasApi"]
 headers = {"Authorization": f"Bearer {canvasToken}"}
 params = {"per_page": 100}
+print()
+bannerTerm = input("Enter Banner Term (e.g., 120255): ")
+print()
 targetDir = '/var/lib/canvas-mgmt/tmp/pdf/'
-rCsvFile = '/var/lib/canvas-mgmt/tmp/canvas-pdf-pages.csv'
-sCsvFile = '/var/lib/canvas-mgmt/tmp/canvas-pdf-list.csv'
+rCsvFile = f'/var/lib/canvas-mgmt/tmp/pdf/{bannerTerm}_canvas-pdf-pages.csv'    #results file
+sCsvFile = f'/var/lib/canvas-mgmt/tmp/pdf/{bannerTerm}_canvas-pdf-list.csv'     #source file
+print(f'targetDir: {targetDir}')
+print(f'rCsvFile:  {rCsvFile}')
+print(f'sCsvFile:  {sCsvFile}')
+print()
 #
 def download_pdf(url, targetFilePath):
     response = requests.get(url, stream=True)
@@ -24,20 +31,24 @@ def download_pdf(url, targetFilePath):
             if chunk:
                 f.write(chunk)
 #
+if not os.path.exists(sCsvFile):
+    print(f"Source file {sCsvFile} does not exist. Exiting...")
+    sys.exit(1)
+#
 with open(rCsvFile, 'w', newline='') as resultsCsvFile:
     fieldnames = ['numPages', 'totalSize', 'uiucCourseID', 'attachmentID', 'pdfName']
     writer = csv.DictWriter(resultsCsvFile, fieldnames=fieldnames)
     writer.writeheader()
     with open(sCsvFile, 'r', newline='') as sourceCsvFile:
-        csvFile = csv.reader(sourceCsvFile)
+        csvFile = csv.DictReader(sourceCsvFile, delimiter='|')
         for row in csvFile:
             print(row)
-            attachmentID = row[2]
-            uiucCourseID = row[0]
-            pdfName = row[6]
-            pdfURL = row[7]
-            totalSize = row[5]
-            targetFilePath = f'{targetDir}-{uiucCourseID}-{attachmentID}.pdf'
+            attachmentID = row['id']
+            uiucCourseID = row['sis_source_id']
+            pdfName = row['display_name']
+            pdfURL = row['download_url']
+            totalSize = row['size_kb']
+            targetFilePath = f'{targetDir}{uiucCourseID}-{attachmentID}.pdf'
             try:
                 r = download_pdf(pdfURL, targetFilePath)
                 with open(targetFilePath, 'rb') as f:
